t3301: test diagnose messages for too few/many paramters
Commit 7d0cbb3af8 ("notes: check number of parameters to "git notes
copy"", 2010-06-28) added a test for too many or too few of
parameters provided to `git notes copy'.

However, the test only ensures that the command will fail but it
doesn't really check if it fails because of number of parameters.

If we accidentally lifted the check inside our code base, the test
may still have failed because the provided parameter is not a valid
ref.

Correct it.

Signed-off-by: Doan Tran Cong Danh <congdanhqx@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>